Reduced dictionary sizes to optimize fitting in processor caches
A dictionary is now taking 65K * 4 = 256K memory, instead of 512K. It is thus easier for a processor to fit it in its lower Ln caches, the only drawback being that cache line invalidation happens twice more often (a dictionary entry takes 4 bytes instead of 8) on entry updates. But overall this seems beneficiary on all platforms tested.